dashboard/streamlit/requirements.txt
streamlit==1.27.2
pandas==2.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
altair==5.1.1
